requirements.txt
absl-py==0.11.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Cython==0.29.21
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.1
idna==2.10
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
pandas==1.2.0
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyparsing==2.4.7
PySimpleGUI==4.34.0
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scipy==1.6.0
seaborn==0.11.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
thop==0.0.31.post2005241907
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
